+++ www/act-map.js
@@ -147,6 +147,7 @@ act('pushSettings', pushSettings);
 act('kbSetKind', kbSetKind);
 act('kbSlot', kbSlot);
 act('kbLtTap', kbLtTap);
+act('kbFontOut', kbFontOut);
 act('langOpen', langOpen);
 act('langDrop', langDrop);
 act('langNew', langNew);